replaced hardcoded 460 focal length with actual focal length in the whole codebase

+++ feature_tracker/src/parameters.h
@@ -4,7 +4,7 @@
 
 extern int ROW;
 extern int COL;
-extern int FOCAL_LENGTH;
+extern double FOCAL_LENGTH;
 const int NUM_OF_CAM = 1;
